Refactor to use game instead of board for logic
