+++ src/App.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
 import './App.css';
-import Board from './components/board';
-import Square from './components/square';
+import Game from './components/game';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
           <h2>The Game of Life</h2>
-          <Board></Board>
+          <Game></Game>
       </div>
     );
   }
